Updates installation sketch
32u2 is also flashable with IDE 1.5.8 or higher now (because 1.0.6
compiles with more flash size)
I hit autoformat, thatswhy there is a lot of change in the files.
I also had to deactivate some of the AVRs since they are not important
to this sketch but take flash. This was needed to suit in the 32u2
bootloader.

Corrected fuses in readme